gcp/projects/cuttlefish-hackathon/hackathon-controller/api/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
google-auth>=2.23.3
google-cloud-firestore>=2.14.0
